# openapi.yaml
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/concepts":{"post":{"summary":"Concepts","operationId":"concepts_concepts_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConceptsPayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Concept"},"title":"Response Concepts Concepts Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/vendors":{"post":{"summary":"Vendors","operationId":"vendors_vendors_post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VendorPayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vendor"},"title":"Response Vendors Vendors Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Concept":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"}},"type":"object","required":["name","description"],"title":"Concept"},"ConceptsPayload":{"properties":{"event_ideas":{"type":"string","title":"Event Ideas"},"event_profile":{"$ref":"#/components/schemas/EventProfile"},"user_feedback":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Feedback","default":""}},"type":"object","required":["event_ideas","event_profile"],"title":"ConceptsPayload"},"EventProfile":{"properties":{"description":{"type":"string","title":"Description"},"min_budget":{"type":"integer","title":"Min Budget"},"max_budget":{"type":"integer","title":"Max Budget"},"min_guests":{"type":"integer","title":"Min Guests"},"max_guests":{"type":"integer","title":"Max Guests"},"event_datetime":{"type":"string","title":"Event Datetime"},"must_haves":{"items":{"type":"string"},"type":"array","title":"Must Haves"},"nice_to_haves":{"items":{"type":"string"},"type":"array","title":"Nice To Haves"},"things_to_avoid":{"items":{"type":"string"},"type":"array","title":"Things To Avoid"},"special_needs":{"type":"string","title":"Special Needs"},"restrictions":{"type":"string","title":"Restrictions"},"concept":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Concept","default":""}},"type":"object","required":["description","min_budget","max_budget","min_guests","max_guests","event_datetime","must_haves","nice_to_haves","things_to_avoid","special_needs","restrictions"],"title":"EventProfile"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Ideas":{"properties":{"name":{"type":"string","title":"Name"},"explanation":{"type":"string","title":"Explanation"}},"type":"object","required":["name","explanation"],"title":"Ideas"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"Vendor":{"properties":{"category":{"type":"string","title":"Category"},"description":{"type":"string","title":"Description"},"ideas":{"items":{"$ref":"#/components/schemas/Ideas"},"type":"array","title":"Ideas"}},"type":"object","required":["category","description","ideas"],"title":"Vendor"},"VendorPayload":{"properties":{"event_profile":{"$ref":"#/components/schemas/EventProfile"},"user_feedback":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Feedback","default":""}},"type":"object","required":["event_profile"],"title":"VendorPayload"}}}}